chore: rename utils file for tailwind
